+++ pkg/analyzer/test/src/summary/elements/type_alias_test.dart
@@ -2629,10 +2629,6 @@ library
 ''');
   }
 
-  @SkippedTest(
-    issue: 'https://github.com/dart-lang/sdk/issues/45291',
-    reason: 'Type dynamic is special, no support for its aliases yet',
-  )
   test_typedef_nonFunction_aliasElement_dynamic() async {
     var library = await buildLibrary(r'''
 typedef A = dynamic;
@@ -2640,8 +2636,35 @@ void f(A a) {}
 ''');
 
     checkElementText(library, r'''
-typedef A = dynamic;
-void f(dynamic<aliasElement: self::@typeAlias::A> a) {}
+library
+  reference: <testLibrary>
+  fragments
+    #F0 <testLibraryFragment>
+      element: <testLibrary>
+      typeAliases
+        #F1 A (nameOffset:8) (firstTokenOffset:0) (offset:8)
+          element: <testLibrary>::@typeAlias::A
+      functions
+        #F2 isOriginDeclaration f (nameOffset:26) (firstTokenOffset:21) (offset:26)
+          element: <testLibrary>::@function::f
+          formalParameters
+            #F3 requiredPositional a (nameOffset:30) (firstTokenOffset:28) (offset:30)
+              element: <testLibrary>::@function::f::@formalParameter::a
+  typeAliases
+    A
+      reference: <testLibrary>::@typeAlias::A
+      firstFragment: #F1
+      aliasedType: dynamic
+  functions
+    isOriginDeclaration f
+      reference: <testLibrary>::@function::f
+      firstFragment: #F2
+      formalParameters
+        #E0 requiredPositional a
+          firstFragment: #F3
+          type: dynamic
+            alias: <testLibrary>::@typeAlias::A
+      returnType: void
 ''');
   }
 
@@ -2788,10 +2811,6 @@ library
 ''');
   }
 
-  @SkippedTest(
-    issue: 'https://github.com/dart-lang/sdk/issues/45291',
-    reason: 'Type Never is special, no support for its aliases yet',
-  )
   test_typedef_nonFunction_aliasElement_never() async {
     var library = await buildLibrary(r'''
 typedef A1 = Never;
@@ -2801,10 +2820,63 @@ void f2(A2<int> a) {}
 ''');
 
     checkElementText(library, r'''
-typedef A1 = Never;
-typedef A2<T> = Never?;
-void f1(Never<aliasElement: self::@typeAlias::A1> a) {}
-void f2(Never?<aliasElement: self::@typeAlias::A2, aliasArguments: [int]> a) {}
+library
+  reference: <testLibrary>
+  fragments
+    #F0 <testLibraryFragment>
+      element: <testLibrary>
+      typeAliases
+        #F1 A1 (nameOffset:8) (firstTokenOffset:0) (offset:8)
+          element: <testLibrary>::@typeAlias::A1
+        #F2 A2 (nameOffset:28) (firstTokenOffset:20) (offset:28)
+          element: <testLibrary>::@typeAlias::A2
+          typeParameters
+            #F3 T (nameOffset:31) (firstTokenOffset:31) (offset:31)
+              element: #E0 T
+      functions
+        #F4 isOriginDeclaration f1 (nameOffset:49) (firstTokenOffset:44) (offset:49)
+          element: <testLibrary>::@function::f1
+          formalParameters
+            #F5 requiredPositional a (nameOffset:55) (firstTokenOffset:52) (offset:55)
+              element: <testLibrary>::@function::f1::@formalParameter::a
+        #F6 isOriginDeclaration f2 (nameOffset:66) (firstTokenOffset:61) (offset:66)
+          element: <testLibrary>::@function::f2
+          formalParameters
+            #F7 requiredPositional a (nameOffset:77) (firstTokenOffset:69) (offset:77)
+              element: <testLibrary>::@function::f2::@formalParameter::a
+  typeAliases
+    A1
+      reference: <testLibrary>::@typeAlias::A1
+      firstFragment: #F1
+      aliasedType: Never
+    A2
+      reference: <testLibrary>::@typeAlias::A2
+      firstFragment: #F2
+      typeParameters
+        #E0 T
+          firstFragment: #F3
+      aliasedType: Never?
+  functions
+    isOriginDeclaration f1
+      reference: <testLibrary>::@function::f1
+      firstFragment: #F4
+      formalParameters
+        #E1 requiredPositional a
+          firstFragment: #F5
+          type: Never
+            alias: <testLibrary>::@typeAlias::A1
+      returnType: void
+    isOriginDeclaration f2
+      reference: <testLibrary>::@function::f2
+      firstFragment: #F6
+      formalParameters
+        #E2 requiredPositional a
+          firstFragment: #F7
+          type: Never?
+            alias: <testLibrary>::@typeAlias::A2
+              typeArguments
+                int
+      returnType: void
 ''');
   }
 
@@ -2913,10 +2985,6 @@ library
 ''');
   }
 
-  @SkippedTest(
-    issue: 'https://github.com/dart-lang/sdk/issues/45291',
-    reason: 'Type void is special, no support for its aliases yet',
-  )
   test_typedef_nonFunction_aliasElement_void() async {
     var library = await buildLibrary(r'''
 typedef A = void;
@@ -2924,8 +2992,35 @@ void f(A a) {}
 ''');
 
     checkElementText(library, r'''
-typedef A = void;
-void f(void<aliasElement: self::@typeAlias::A> a) {}
+library
+  reference: <testLibrary>
+  fragments
+    #F0 <testLibraryFragment>
+      element: <testLibrary>
+      typeAliases
+        #F1 A (nameOffset:8) (firstTokenOffset:0) (offset:8)
+          element: <testLibrary>::@typeAlias::A
+      functions
+        #F2 isOriginDeclaration f (nameOffset:23) (firstTokenOffset:18) (offset:23)
+          element: <testLibrary>::@function::f
+          formalParameters
+            #F3 requiredPositional a (nameOffset:27) (firstTokenOffset:25) (offset:27)
+              element: <testLibrary>::@function::f::@formalParameter::a
+  typeAliases
+    A
+      reference: <testLibrary>::@typeAlias::A
+      firstFragment: #F1
+      aliasedType: void
+  functions
+    isOriginDeclaration f
+      reference: <testLibrary>::@function::f
+      firstFragment: #F2
+      formalParameters
+        #E0 requiredPositional a
+          firstFragment: #F3
+          type: void
+            alias: <testLibrary>::@typeAlias::A
+      returnType: void
 ''');
   }
 
@@ -4089,6 +4184,7 @@ library
         #E0 requiredPositional a
           firstFragment: #F3
           type: dynamic
+            alias: <testLibrary>::@typeAlias::A
       returnType: void
 ''');
   }
@@ -4288,6 +4384,7 @@ library
         #E0 requiredPositional a
           firstFragment: #F3
           type: Never
+            alias: <testLibrary>::@typeAlias::A
       returnType: void
 ''');
   }
@@ -4325,6 +4422,7 @@ library
         #E0 requiredPositional a
           firstFragment: #F3
           type: Never?
+            alias: <testLibrary>::@typeAlias::A
       returnType: void
 ''');
   }
@@ -4466,6 +4564,9 @@ library
         #E1 requiredPositional a
           firstFragment: #F4
           type: dynamic
+            alias: <testLibrary>::@typeAlias::A
+              typeArguments
+                dynamic
       returnType: void
     isOriginDeclaration f2
       reference: <testLibrary>::@function::f2
@@ -4532,6 +4633,9 @@ library
         #E1 requiredPositional a
           firstFragment: #F4
           type: dynamic
+            alias: <testLibrary>::@typeAlias::A
+              typeArguments
+                dynamic
       returnType: void
     isOriginDeclaration f2
       reference: <testLibrary>::@function::f2
@@ -4592,6 +4696,7 @@ library
         #E0 requiredPositional a
           firstFragment: #F3
           type: void
+            alias: <testLibrary>::@typeAlias::A
       returnType: void
 ''');
   }